Assign and commit semantics for `NewspaperWorks::Data::WorkDerivatives` WIP [skip ci].

+++ lib/newspaper_works/data/work_derivatives.rb
@@ -2,8 +2,17 @@ require 'hyrax'
 
 module NewspaperWorks
   module Data
+    # Disable below metric, for now, not splitting this class into modules,
+    #   which would just make for more complexity and hinder readabilty.
+    #   TODO: consider compositional refactoring (not mixins), but this
+    #         may make readability/comprehendability higher, and yield
+    #         higher applied/practical complexity.
+    # rubocop:disable Metrics/ClassLength
     class WorkDerivatives
       include NewspaperWorks::Data::FilesetHelper
+      include NewspaperWorks::Data::PathHelper
+
+      attr_accessor :work, :fileset
 
       # mapping of special names Hyrax uses for derivatives, not extension:
       @remap_names = {
@@ -13,19 +22,53 @@ module NewspaperWorks
         attr_accessor :remap_names
       end
 
-      def initialize(work)
-        # context usually work, may be FileSet, may be string id of FileSet
+      def initialize(work, fileset: nil)
+        # adapted context usually work, may be string id of FileSet
         @work = work
+        @fileset = fileset.nil? ? first_fileset : fileset
         # computed name-to-path mapping, initially nil as sentinel for JIT load
         @paths = nil
+        # assignments for attachment
+        @assigned = []
+        # un-assignments for deletion
+        @unassigned = []
+      end
+
+      def state
+        return 'empty' if @assigned.empty? && @paths.keys.empty?
+        return 'dirty' unless @assigned.empty?
+        'saved'
+      end
+
+      def assign(path)
+        path = normalize_path(path)
+        validate_path(path)
+        @assigned.push(path)
+      end
+
+      def unassign(name)
+        # if name is queued path, remove from @assigned queue:
+        @assigned.delete(name) if @assigned.include?(name)
+        # if name is known destination name, remove
+        @unassigned.push(name) if exist?(name)
+      end
+
+      def commit!
+        @unassigned.each { |name| delete(name) }
+        @assigned.each do |path|
+          attach(path, path_destination_name(path))
+        end
+        # reset queues after work is complete
+        @assigned = []
+        @unassigned = []
       end
 
       def attach(file, name)
         mkdir_pairtree
-        path = path_factory.derivative_path_for_reference(work_fileset, name)
+        path = path_factory.derivative_path_for_reference(fileset, name)
         # if file argument is path, copy file
         if file.class == String
-          FileUtils.copy(file, path) if file.class == String
+          FileUtils.copy(file, path)
         else
           # otherwise, presume file is an IO, read, write it
           #   note: does not close input file/IO, presume that is caller's
@@ -40,10 +83,10 @@ module NewspaperWorks
       end
 
       def delete(name, force: nil)
-        path = path_factory.derivative_path_for_reference(work_fileset, name)
+        path = path_factory.derivative_path_for_reference(fileset, name)
         # will remove file, if it exists; won't remove pairtree, even
         #   if it becomes empty, as that is excess scope.
-        FileUtils.rm(path, force: force)
+        FileUtils.rm(path, force: force) if File.exist?(path)
         # finally, reload @paths after mutation
         load_paths
       end
@@ -120,14 +163,14 @@ module NewspaperWorks
 
         # make shared path for derivatives to live, given
         def mkdir_pairtree
-          ensure_fileset_exists
           # Hyrax::DerivativePath has no public method to directly get the
           #   bare pairtree path for derivatives for a fileset, but we
           #   can infer it...
-          path = path_factory.derivative_path_for_reference(work_fileset, '')
+          path = path_factory.derivative_path_for_reference(fileset, '')
           dir = File.join(path.split('/')[0..-2])
           FileUtils.mkdir_p(dir) unless Dir.exist?(dir)
         end
     end
+    # rubocop:enable Metrics/ClassLength
   end
 end
